# openapi.yaml
openapi: 3.1.0
info:
  title: pr-lens-demo API
  version: 1.0.0
  description: |
    Public API for the demo SaaS project-management application. Used as a
    reference repository by PR Lens to exercise its API-change pillar.
servers:
  - url: https://pr-lens-demo.vercel.app
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
  /api/v0/workspaces:
    get:
      tags: [legacy]
      summary: List the calling token's workspace (legacy v0)
      operationId: legacyListWorkspaces
      deprecated: true
      parameters:
        - name: x-api-token
          in: header
          required: true
          schema: { type: string }
      responses:
        "200":
          description: A single-item array with the workspace
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LegacyWorkspace"
  /api/v0/tasks:
    get:
      tags: [legacy]
      summary: List tasks in the calling token's workspace (legacy v0)
      operationId: legacyListTasks
      deprecated: true
      parameters:
        - name: x-api-token
          in: header
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Tasks in the workspace
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LegacyTask"
...
  /api/v1/tasks:
    get:
      tags: [tasks]
      summary: List tasks, optionally filtered by workspace
      operationId: listTasks
      parameters:
        - name: workspaceId
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: A list of tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
    post:
      tags: [tasks]
      summary: Create a task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspaceId, title]
              properties:
                workspaceId: { type: string }
                title: { type: string }
                description: { type: string }
                priority:
                  type: string
                  enum: [none, low, medium, high, urgent]
                assigneeId: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
components:
  schemas:
    LegacyWorkspace:
      type: object
      required: [id, slug, name, plan]
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
        plan: { type: string }
    LegacyTask:
      type: object
      required: [id, workspaceId, title, status]
      properties:
        id: { type: string }
        workspaceId: { type: string }
        title: { type: string }
        status: { type: string }
        assigneeId: { type: string, nullable: true }
...
    Task:
      type: object
      required: [id, workspaceId, title, status, createdAt]
      properties:
        id: { type: string }
        workspaceId: { type: string }
        title: { type: string }
        description: { type: string }
        status:
          type: string
          enum: [todo, in_progress, done, archived]
        priority:
          type: string
          enum: [none, low, medium, high, urgent]
          nullable: true
        assigneeId: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }